:recycle: Move constants that are specific to each folders
